+++ CHANGES.md
@@ -4,6 +4,10 @@ Added a schema compile option, `:promote-nils-to-empty-list?`, which converts
 nils resolved from list fields back into empty lists (use only if you
 have existing clients that break when they see those nils).
 
+The `com.walmartlabs.schema/coercion-failure` function has been deprecated,
+scalar parse and serialize functions can simply throw an exception rather
+than invoke `coercion-failure`.
+
 ## 0.31.0 -- 4 Jan 2019
 
 This release, regrettably, has some backwards incompatible changes (notice the `0.` at the front